Rename future to preview (pt. 1)
