+++ src/cocotb/_init.py
@@ -18,9 +18,9 @@ from typing import cast
 import cocotb
 import cocotb._profiling
 import cocotb._shutdown
-import cocotb.future
 import cocotb.handle
 import cocotb.logging
+import cocotb.preview
 import cocotb.simtime
 import cocotb.simulator
 from cocotb_tools import _env
@@ -32,7 +32,7 @@ def init_package_from_simulation() -> None:
     """Initialize the cocotb package from a simulation context."""
 
     # Initialize subsystems
-    cocotb.future._init()
+    cocotb.preview._init()
     cocotb._shutdown._init()
     cocotb.logging._init()
     cocotb._profiling._init()
